Linear pool - updated to BigInt maths.
