+++ src/pools/linearPool/linearPool.ts
@@ -4,20 +4,6 @@ import { BigNumber as OldBigNumber } from '../../utils/bignumber';
 import { WeiPerEther as ONE } from '@ethersproject/constants';
 import { isSameAddress } from '../../utils';
 import {
-    _calcBptInPerWrappedOut,
-    _calcMainOutPerWrappedIn,
-    _calcWrappedOutPerMainIn,
-    _calcBptOutPerMainIn,
-    _calcMainOutPerBptIn,
-    _calcBptOutPerWrappedIn,
-    _calcWrappedOutPerBptIn,
-    _calcWrappedInPerMainOut,
-    _calcMainInPerWrappedOut,
-    _calcMainInPerBptOut,
-    _calcBptInPerMainOut,
-    _calcWrappedInPerBptOut,
-} from './exactMaths';
-import {
     PoolBase,
     PoolTypes,
     SwapPairType,
@@ -27,10 +13,18 @@ import {
     SubgraphToken,
 } from '../../types';
 import {
-    _exactMainTokenInForBPTOut,
-    _exactBPTInForMainTokenOut,
-    _mainTokenInForExactBPTOut,
-    _BPTInForExactMainTokenOut,
+    _calcBptOutPerMainIn,
+    _calcBptInPerWrappedOut,
+    _calcBptInPerMainOut,
+    _calcWrappedOutPerMainIn,
+    _calcWrappedInPerMainOut,
+    _calcMainInPerBptOut,
+    _calcMainOutPerBptIn,
+    _calcMainOutPerWrappedIn,
+    _calcMainInPerWrappedOut,
+    _calcBptOutPerWrappedIn,
+    _calcWrappedInPerBptOut,
+    _calcWrappedOutPerBptIn,
     _spotPriceAfterSwapExactTokenInForTokenOut,
     _spotPriceAfterSwapExactTokenInForBPTOut,
     _spotPriceAfterSwapExactBPTInForTokenOut,
@@ -62,12 +56,12 @@ type LinearPoolToken = Pick<
 export type LinearPoolPairData = PoolPairBase & {
     pairType: PairTypes;
     wrappedBalance: OldBigNumber; // If main token is USDC then wrapped token is aUSDC (or a wrapped version of it)
+    wrappedBalanceScaled: BigNumber; // If main token is USDC then wrapped token is aUSDC (or a wrapped version of it)
     wrappedDecimals: number;
-    rate: OldBigNumber; // PriceRate of wrapped token
+    rate: BigNumber; // PriceRate of wrapped token
     lowerTarget: BigNumber; // Target determine the range where there are positive, zero or negative fees
     upperTarget: BigNumber; // when the "main token" has a balance below lowerTarget, there are negative fees when adding main token
     mainBalanceScaled: BigNumber; // Scaled are used for EVM/SDK maths
-    wrappedBalanceScaled: BigNumber;
     bptBalanceScaled: BigNumber;
     virtualBptSupply: BigNumber;
 };
@@ -178,15 +172,10 @@ export class LinearPool implements PoolBase {
         const allBalancesScaled = this.tokens.map(({ balance }) =>
             parseFixed(balance, 18)
         );
-        const priceRate = this.tokens[this.wrappedIndex].priceRate;
-        const mainBalanceScaled = allBalancesScaled[this.mainIndex];
-        const wrappedBalanceScaled = allBalancesScaled[this.wrappedIndex]
-            .mul(parseFixed(priceRate, 18))
-            .div(ONE);
-        const bptBalanceScaled = allBalancesScaled[this.bptIndex];
         // https://github.com/balancer-labs/balancer-v2-monorepo/blob/88a14eb623f6a22ef3f1afc5a8c49ebfa7eeceed/pkg/pool-linear/contracts/LinearPool.sol#L247
         // VirtualBPTSupply must be used for the maths
         // TO DO - SG should be updated to so that totalShares should return VirtualSupply
+        const bptBalanceScaled = allBalancesScaled[this.bptIndex];
         const virtualBptSupply = this.MAX_TOKEN_BALANCE.sub(bptBalanceScaled);
 
         const poolPairData: LinearPoolPairData = {
@@ -205,12 +194,12 @@ export class LinearPool implements PoolBase {
                 bnum(this.tokens[this.wrappedIndex].balance),
                 this.wrappedDecimals
             ),
+            wrappedBalanceScaled: allBalancesScaled[this.wrappedIndex], // Note this is not multiplied by rate
             wrappedDecimals: this.wrappedDecimals,
-            rate: scale(bnum(priceRate), 18),
+            rate: parseFixed(this.tokens[this.wrappedIndex].priceRate, 18),
             lowerTarget: this.lowerTarget,
             upperTarget: this.upperTarget,
-            mainBalanceScaled,
-            wrappedBalanceScaled,
+            mainBalanceScaled: allBalancesScaled[this.mainIndex],
             bptBalanceScaled,
             virtualBptSupply,
         };
@@ -223,7 +212,7 @@ export class LinearPool implements PoolBase {
     }
 
     getLimitAmountSwap(
-        poolPairData: PoolPairBase,
+        poolPairData: LinearPoolPairData,
         swapType: SwapTypes
     ): OldBigNumber {
         // Needs to return human scaled numbers
@@ -235,12 +224,13 @@ export class LinearPool implements PoolBase {
 
         if (swapType === SwapTypes.SwapExactIn) {
             if (linearPoolPairData.pairType === PairTypes.MainTokenToBpt) {
-                return _mainTokenInForExactBPTOut(
-                    balanceOutHuman,
-                    linearPoolPairData
-                )
-                    .times(bnum(this.ALMOST_ONE.toString()))
-                    .div(bnum(ONE.toString()));
+                return this._mainTokenInForExactBPTOut(
+                    poolPairData,
+                    balanceOutHuman
+                        .times(this.ALMOST_ONE.toString())
+                        .div(ONE.toString()),
+                    true
+                );
             } else if (
                 linearPoolPairData.pairType === PairTypes.WrappedTokenToBpt
             ) {
@@ -251,37 +241,23 @@ export class LinearPool implements PoolBase {
             ) {
                 // Limit is amount of BPT in for pool balance of tokenOut
                 // Amount must be in human scale
-                const limit = _BPTInForExactMainTokenOut(
-                    balanceOutHuman,
-                    linearPoolPairData
-                )
-                    .times(bnum(this.ALMOST_ONE.toString()))
-                    .div(bnum(ONE.toString()));
-
-                return limit;
+                return this._BPTInForExactMainTokenOut(
+                    linearPoolPairData,
+                    balanceOutHuman
+                        .times(this.ALMOST_ONE.toString())
+                        .div(ONE.toString()),
+                    true
+                );
             } else if (
                 linearPoolPairData.pairType === PairTypes.BptToWrappedToken
             ) {
-                // Limit is amount of BPT in for pool balance of tokenOut
-                const limit = _calcBptInPerWrappedOut(
-                    bnum(poolPairData.balanceOut.toString()),
-                    bnum(linearPoolPairData.mainBalanceScaled.toString()),
-                    bnum(linearPoolPairData.wrappedBalanceScaled.toString()),
-                    bnum(linearPoolPairData.bptBalanceScaled.toString()),
-                    {
-                        fee: bnum(poolPairData.swapFee.toString()),
-                        lowerTarget: bnum(
-                            linearPoolPairData.lowerTarget.toString()
-                        ),
-                        upperTarget: bnum(
-                            linearPoolPairData.upperTarget.toString()
-                        ),
-                    }
-                )
-                    .times(bnum(this.ALMOST_ONE.toString()))
-                    .div(bnum(ONE.toString()))
-                    .div(bnum(ONE.toString()));
-
+                const limit = this._BPTInForExactWrappedTokenOut(
+                    poolPairData,
+                    balanceOutHuman
+                        .times(this.ALMOST_ONE.toString())
+                        .div(ONE.toString()),
+                    true
+                );
                 // Returning Human scale
                 return limit;
             } else if (
@@ -384,22 +360,26 @@ export class LinearPool implements PoolBase {
         try {
             // All values should use 1e18 fixed point
             // i.e. 1USDC => 1e18 not 1e6
-            const amtScaled = scale(amount, 18);
+            const amtScaled = parseFixed(amount.toString(), 18);
 
             const amt = _calcMainOutPerWrappedIn(
-                amtScaled,
-                bnum(poolPairData.mainBalanceScaled.toString()),
+                amtScaled.toBigInt(),
+                poolPairData.mainBalanceScaled.toBigInt(),
                 {
-                    fee: bnum(poolPairData.swapFee.toString()),
-                    lowerTarget: bnum(poolPairData.lowerTarget.toString()),
-                    upperTarget: bnum(poolPairData.upperTarget.toString()),
+                    fee: poolPairData.swapFee.toBigInt(),
+                    lowerTarget: poolPairData.lowerTarget.toBigInt(),
+                    upperTarget: poolPairData.upperTarget.toBigInt(),
+                    rate: poolPairData.rate.toBigInt(),
                 }
             );
             // return human readable number
             // Using BigNumber.js decimalPlaces (dp), allows us to consider token decimal accuracy correctly,
             // i.e. when using token with 2decimals 0.002 should be returned as 0
             // Uses ROUND_DOWN mode (1)
-            return scale(amt, -18).dp(poolPairData.decimalsOut, 1);
+            return scale(bnum(amt.toString()), -18).dp(
+                poolPairData.decimalsOut,
+                1
+            );
         } catch (err) {
             return ZERO;
         }
@@ -413,22 +393,26 @@ export class LinearPool implements PoolBase {
         try {
             // All values should use 1e18 fixed point
             // i.e. 1USDC => 1e18 not 1e6
-            const amtScaled = scale(amount, 18);
+            const amtScaled = parseFixed(amount.toString(), 18);
 
             const amt = _calcWrappedOutPerMainIn(
-                amtScaled,
-                bnum(poolPairData.mainBalanceScaled.toString()),
+                amtScaled.toBigInt(),
+                poolPairData.mainBalanceScaled.toBigInt(),
                 {
-                    fee: bnum(poolPairData.swapFee.toString()),
-                    lowerTarget: bnum(poolPairData.lowerTarget.toString()),
-                    upperTarget: bnum(poolPairData.upperTarget.toString()),
+                    fee: poolPairData.swapFee.toBigInt(),
+                    lowerTarget: poolPairData.lowerTarget.toBigInt(),
+                    upperTarget: poolPairData.upperTarget.toBigInt(),
+                    rate: poolPairData.rate.toBigInt(),
                 }
             );
             // return human readable number
             // Using BigNumber.js decimalPlaces (dp), allows us to consider token decimal accuracy correctly,
             // i.e. when using token with 2decimals 0.002 should be returned as 0
             // Uses ROUND_DOWN mode (1)
-            return scale(amt, -18).dp(poolPairData.decimalsOut, 1);
+            return scale(bnum(amt.toString()), -18).dp(
+                poolPairData.decimalsOut,
+                1
+            );
         } catch (err) {
             return ZERO;
         }
@@ -439,33 +423,33 @@ export class LinearPool implements PoolBase {
         amount: OldBigNumber,
         exact: boolean
     ): OldBigNumber {
-        if (exact) {
-            try {
-                // All values should use 1e18 fixed point
-                // i.e. 1USDC => 1e18 not 1e6
-                const amtScaled = scale(amount, 18);
+        try {
+            // All values should use 1e18 fixed point
+            // i.e. 1USDC => 1e18 not 1e6
+            const amtScaled = parseFixed(amount.toString(), 18);
 
-                const amt = _calcBptOutPerMainIn(
-                    amtScaled,
-                    bnum(poolPairData.mainBalanceScaled.toString()),
-                    bnum(poolPairData.wrappedBalanceScaled.toString()),
-                    bnum(poolPairData.virtualBptSupply.toString()),
-                    {
-                        fee: bnum(poolPairData.swapFee.toString()),
-                        lowerTarget: bnum(poolPairData.lowerTarget.toString()),
-                        upperTarget: bnum(poolPairData.upperTarget.toString()),
-                    }
-                );
-                // return human readable number
-                // Using BigNumber.js decimalPlaces (dp), allows us to consider token decimal accuracy correctly,
-                // i.e. when using token with 2decimals 0.002 should be returned as 0
-                // Uses ROUND_DOWN mode (1)
-                return scale(amt, -18).dp(poolPairData.decimalsOut, 1);
-            } catch (err) {
-                return ZERO;
-            }
-        } else {
-            return _exactMainTokenInForBPTOut(amount, poolPairData);
+            const amt = _calcBptOutPerMainIn(
+                amtScaled.toBigInt(),
+                poolPairData.mainBalanceScaled.toBigInt(),
+                poolPairData.wrappedBalanceScaled.toBigInt(),
+                poolPairData.virtualBptSupply.toBigInt(),
+                {
+                    fee: poolPairData.swapFee.toBigInt(),
+                    lowerTarget: poolPairData.lowerTarget.toBigInt(),
+                    upperTarget: poolPairData.upperTarget.toBigInt(),
+                    rate: poolPairData.rate.toBigInt(),
+                }
+            );
+            // return human readable number
+            // Using BigNumber.js decimalPlaces (dp), allows us to consider token decimal accuracy correctly,
+            // i.e. when using token with 2decimals 0.002 should be returned as 0
+            // Uses ROUND_DOWN mode (1)
+            return scale(bnum(amt.toString()), -18).dp(
+                poolPairData.decimalsOut,
+                1
+            );
+        } catch (err) {
+            return ZERO;
         }
     }
 
@@ -474,32 +458,34 @@ export class LinearPool implements PoolBase {
         amount: OldBigNumber,
         exact: boolean
     ): OldBigNumber {
-        if (exact) {
-            try {
-                // All values should use 1e18 fixed point
-                // i.e. 1USDC => 1e18 not 1e6
-                const amtScaled = scale(amount, 18);
+        try {
+            // All values should use 1e18 fixed point
+            // i.e. 1USDC => 1e18 not 1e6
+            const amtScaled = parseFixed(amount.toString(), 18);
 
-                const amt = _calcMainOutPerBptIn(
-                    amtScaled,
-                    bnum(poolPairData.mainBalanceScaled.toString()),
-                    bnum(poolPairData.wrappedBalanceScaled.toString()),
-                    bnum(poolPairData.virtualBptSupply.toString()),
-                    {
-                        fee: bnum(poolPairData.swapFee.toString()),
-                        lowerTarget: bnum(poolPairData.lowerTarget.toString()),
-                        upperTarget: bnum(poolPairData.upperTarget.toString()),
-                    }
-                );
-                // return human readable number
-                // Using BigNumber.js decimalPlaces (dp), allows us to consider token decimal accuracy correctly,
-                // i.e. when using token with 2decimals 0.002 should be returned as 0
-                // Uses ROUND_DOWN mode (1)
-                return scale(amt, -18).dp(poolPairData.decimalsOut, 1);
-            } catch (err) {
-                return ZERO;
-            }
-        } else return _exactBPTInForMainTokenOut(amount, poolPairData);
+            const amt = _calcMainOutPerBptIn(
+                amtScaled.toBigInt(),
+                poolPairData.mainBalanceScaled.toBigInt(),
+                poolPairData.wrappedBalanceScaled.toBigInt(),
+                poolPairData.virtualBptSupply.toBigInt(),
+                {
+                    fee: poolPairData.swapFee.toBigInt(),
+                    lowerTarget: poolPairData.lowerTarget.toBigInt(),
+                    upperTarget: poolPairData.upperTarget.toBigInt(),
+                    rate: poolPairData.rate.toBigInt(),
+                }
+            );
+            // return human readable number
+            // Using BigNumber.js decimalPlaces (dp), allows us to consider token decimal accuracy correctly,
+            // i.e. when using token with 2decimals 0.002 should be returned as 0
+            // Uses ROUND_DOWN mode (1)
+            return scale(bnum(amt.toString()), -18).dp(
+                poolPairData.decimalsOut,
+                1
+            );
+        } catch (err) {
+            return ZERO;
+        }
     }
 
     _exactWrappedTokenInForBPTOut(
@@ -510,24 +496,30 @@ export class LinearPool implements PoolBase {
         try {
             // All values should use 1e18 fixed point
             // i.e. 1USDC => 1e18 not 1e6
-            const amtScaled = scale(amount, 18);
+            const amtNoRate = parseFixed(amount.toString(), 18)
+                .mul(ONE)
+                .div(poolPairData.rate);
 
             const amt = _calcBptOutPerWrappedIn(
-                amtScaled,
-                bnum(poolPairData.mainBalanceScaled.toString()),
-                bnum(poolPairData.wrappedBalanceScaled.toString()),
-                bnum(poolPairData.virtualBptSupply.toString()),
+                amtNoRate.toBigInt(),
+                poolPairData.mainBalanceScaled.toBigInt(),
+                poolPairData.wrappedBalanceScaled.toBigInt(),
+                poolPairData.virtualBptSupply.toBigInt(),
                 {
-                    fee: bnum(poolPairData.swapFee.toString()),
-                    lowerTarget: bnum(poolPairData.lowerTarget.toString()),
-                    upperTarget: bnum(poolPairData.upperTarget.toString()),
+                    fee: poolPairData.swapFee.toBigInt(),
+                    lowerTarget: poolPairData.lowerTarget.toBigInt(),
+                    upperTarget: poolPairData.upperTarget.toBigInt(),
+                    rate: poolPairData.rate.toBigInt(),
                 }
             );
             // return human readable number
             // Using BigNumber.js decimalPlaces (dp), allows us to consider token decimal accuracy correctly,
             // i.e. when using token with 2decimals 0.002 should be returned as 0
             // Uses ROUND_DOWN mode (1)
-            return scale(amt, -18).dp(poolPairData.decimalsOut, 1);
+            return scale(bnum(amt.toString()), -18).dp(
+                poolPairData.decimalsOut,
+                1
+            );
         } catch (err) {
             return ZERO;
         }
@@ -541,24 +533,32 @@ export class LinearPool implements PoolBase {
         try {
             // All values should use 1e18 fixed point
             // i.e. 1USDC => 1e18 not 1e6
-            const amtScaled = scale(amount, 18);
+            const amtScaled = parseFixed(amount.toString(), 18);
 
             const amt = _calcWrappedOutPerBptIn(
-                amtScaled,
-                bnum(poolPairData.mainBalanceScaled.toString()),
-                bnum(poolPairData.wrappedBalanceScaled.toString()),
-                bnum(poolPairData.virtualBptSupply.toString()),
+                amtScaled.toBigInt(),
+                poolPairData.mainBalanceScaled.toBigInt(),
+                poolPairData.wrappedBalanceScaled.toBigInt(),
+                poolPairData.virtualBptSupply.toBigInt(),
                 {
-                    fee: bnum(poolPairData.swapFee.toString()),
-                    lowerTarget: bnum(poolPairData.lowerTarget.toString()),
-                    upperTarget: bnum(poolPairData.upperTarget.toString()),
+                    fee: poolPairData.swapFee.toBigInt(),
+                    lowerTarget: poolPairData.lowerTarget.toBigInt(),
+                    upperTarget: poolPairData.upperTarget.toBigInt(),
+                    rate: poolPairData.rate.toBigInt(),
                 }
             );
+
+            const amtWithRate = BigNumber.from(amt)
+                .mul(poolPairData.rate)
+                .div(ONE);
             // return human readable number
             // Using BigNumber.js decimalPlaces (dp), allows us to consider token decimal accuracy correctly,
             // i.e. when using token with 2decimals 0.002 should be returned as 0
             // Uses ROUND_DOWN mode (1)
-            return scale(amt, -18).dp(poolPairData.decimalsOut, 1);
+            return scale(bnum(amtWithRate.toString()), -18).dp(
+                poolPairData.decimalsOut,
+                1
+            );
         } catch (err) {
             return ZERO;
         }
@@ -612,22 +612,26 @@ export class LinearPool implements PoolBase {
         try {
             // All values should use 1e18 fixed point
             // i.e. 1USDC => 1e18 not 1e6
-            const amtScaled = scale(amount, 18);
+            const amtScaled = parseFixed(amount.toString(), 18);
 
             const amt = _calcWrappedInPerMainOut(
-                amtScaled,
-                bnum(poolPairData.mainBalanceScaled.toString()),
+                amtScaled.toBigInt(),
+                poolPairData.mainBalanceScaled.toBigInt(),
                 {
-                    fee: bnum(poolPairData.swapFee.toString()),
-                    lowerTarget: bnum(poolPairData.lowerTarget.toString()),
-                    upperTarget: bnum(poolPairData.upperTarget.toString()),
+                    fee: poolPairData.swapFee.toBigInt(),
+                    lowerTarget: poolPairData.lowerTarget.toBigInt(),
+                    upperTarget: poolPairData.upperTarget.toBigInt(),
+                    rate: poolPairData.rate.toBigInt(),
                 }
             );
             // return human readable number
             // Using BigNumber.js decimalPlaces (dp), allows us to consider token decimal accuracy correctly,
             // i.e. when using token with 2decimals 0.002 should be returned as 0
             // Uses ROUND_DOWN mode (1)
-            return scale(amt, -18).dp(poolPairData.decimalsOut, 1);
+            return scale(bnum(amt.toString()), -18).dp(
+                poolPairData.decimalsOut,
+                1
+            );
         } catch (err) {
             return ZERO;
         }
@@ -641,22 +645,26 @@ export class LinearPool implements PoolBase {
         try {
             // All values should use 1e18 fixed point
             // i.e. 1USDC => 1e18 not 1e6
-            const amtScaled = scale(amount, 18);
+            const amtScaled = parseFixed(amount.toString(), 18);
 
             const amt = _calcMainInPerWrappedOut(
-                amtScaled,
-                bnum(poolPairData.mainBalanceScaled.toString()),
+                amtScaled.toBigInt(),
+                poolPairData.mainBalanceScaled.toBigInt(),
                 {
-                    fee: bnum(poolPairData.swapFee.toString()),
-                    lowerTarget: bnum(poolPairData.lowerTarget.toString()),
-                    upperTarget: bnum(poolPairData.upperTarget.toString()),
+                    fee: poolPairData.swapFee.toBigInt(),
+                    lowerTarget: poolPairData.lowerTarget.toBigInt(),
+                    upperTarget: poolPairData.upperTarget.toBigInt(),
+                    rate: poolPairData.rate.toBigInt(),
                 }
             );
             // return human readable number
             // Using BigNumber.js decimalPlaces (dp), allows us to consider token decimal accuracy correctly,
             // i.e. when using token with 2decimals 0.002 should be returned as 0
             // Uses ROUND_DOWN mode (1)
-            return scale(amt, -18).dp(poolPairData.decimalsOut, 1);
+            return scale(bnum(amt.toString()), -18).dp(
+                poolPairData.decimalsOut,
+                1
+            );
         } catch (err) {
             return ZERO;
         }
@@ -667,34 +675,35 @@ export class LinearPool implements PoolBase {
         amount: OldBigNumber,
         exact: boolean
     ): OldBigNumber {
-        if (exact) {
-            try {
-                // All values should use 1e18 fixed point
-                // i.e. 1USDC => 1e18 not 1e6
-                const amtScaled = scale(amount, 18);
-                // in = main
-                // out = BPT
-                const amt = _calcMainInPerBptOut(
-                    amtScaled,
-                    bnum(poolPairData.mainBalanceScaled.toString()),
-                    bnum(poolPairData.wrappedBalanceScaled.toString()),
-                    bnum(poolPairData.virtualBptSupply.toString()),
-                    {
-                        fee: bnum(poolPairData.swapFee.toString()),
-                        lowerTarget: bnum(poolPairData.lowerTarget.toString()),
-                        upperTarget: bnum(poolPairData.upperTarget.toString()),
-                    }
-                );
-                // return human readable number
-                // Using BigNumber.js decimalPlaces (dp), allows us to consider token decimal accuracy correctly,
-                // i.e. when using token with 2decimals 0.002 should be returned as 0
-                // Uses ROUND_UP mode (0)
-                return scale(amt, -18).dp(poolPairData.decimalsIn, 0);
-            } catch (err) {
-                return ZERO;
-            }
+        try {
+            // All values should use 1e18 fixed point
+            // i.e. 1USDC => 1e18 not 1e6
+            const amtScaled = parseFixed(amount.toString(), 18);
+            // in = main
+            // out = BPT
+            const amt = _calcMainInPerBptOut(
+                amtScaled.toBigInt(),
+                poolPairData.mainBalanceScaled.toBigInt(),
+                poolPairData.wrappedBalanceScaled.toBigInt(),
+                poolPairData.virtualBptSupply.toBigInt(),
+                {
+                    fee: poolPairData.swapFee.toBigInt(),
+                    lowerTarget: poolPairData.lowerTarget.toBigInt(),
+                    upperTarget: poolPairData.upperTarget.toBigInt(),
+                    rate: poolPairData.rate.toBigInt(),
+                }
+            );
+            // return human readable number
+            // Using BigNumber.js decimalPlaces (dp), allows us to consider token decimal accuracy correctly,
+            // i.e. when using token with 2decimals 0.002 should be returned as 0
+            // Uses ROUND_UP mode (0)
+            return scale(bnum(amt.toString()), -18).dp(
+                poolPairData.decimalsIn,
+                0
+            );
+        } catch (err) {
+            return ZERO;
         }
-        return _mainTokenInForExactBPTOut(amount, poolPairData);
     }
 
     _BPTInForExactMainTokenOut(
@@ -702,33 +711,35 @@ export class LinearPool implements PoolBase {
         amount: OldBigNumber,
         exact: boolean
     ): OldBigNumber {
-        if (exact) {
-            try {
-                // All values should use 1e18 fixed point
-                // i.e. 1USDC => 1e18 not 1e6
-                const amtScaled = scale(amount, 18);
+        try {
+            // All values should use 1e18 fixed point
+            // i.e. 1USDC => 1e18 not 1e6
+            const amtScaled = parseFixed(amount.toString(), 18);
 
-                const amt = _calcBptInPerMainOut(
-                    amtScaled,
-                    bnum(poolPairData.mainBalanceScaled.toString()),
-                    bnum(poolPairData.wrappedBalanceScaled.toString()),
-                    bnum(poolPairData.virtualBptSupply.toString()),
-                    {
-                        fee: bnum(poolPairData.swapFee.toString()),
-                        lowerTarget: bnum(poolPairData.lowerTarget.toString()),
-                        upperTarget: bnum(poolPairData.upperTarget.toString()),
-                    }
-                );
-                // return human readable number
-                // Using BigNumber.js decimalPlaces (dp), allows us to consider token decimal accuracy correctly,
-                // i.e. when using token with 2decimals 0.002 should be returned as 0
-                // Uses ROUND_UP mode (0)
-                return scale(amt, -18).dp(poolPairData.decimalsIn, 0);
-            } catch (err) {
-                return ZERO;
-            }
+            const amt = _calcBptInPerMainOut(
+                amtScaled.toBigInt(),
+                poolPairData.mainBalanceScaled.toBigInt(),
+                poolPairData.wrappedBalanceScaled.toBigInt(),
+                poolPairData.virtualBptSupply.toBigInt(),
+                {
+                    fee: poolPairData.swapFee.toBigInt(),
+                    lowerTarget: poolPairData.lowerTarget.toBigInt(),
+                    upperTarget: poolPairData.upperTarget.toBigInt(),
+                    rate: poolPairData.rate.toBigInt(),
+                }
+            );
+            // return human readable number
+            // Using BigNumber.js decimalPlaces (dp), allows us to consider token decimal accuracy correctly,
+            // i.e. when using token with 2decimals 0.002 should be returned as 0
+            // Uses ROUND_UP mode (0)
+            return scale(bnum(amt.toString()), -18).dp(
+                poolPairData.decimalsIn,
+                0
+            );
+        } catch (err) {
+            console.log('OUCH?', err);
+            return ZERO;
         }
-        return _BPTInForExactMainTokenOut(amount, poolPairData);
     }
 
     _wrappedTokenInForExactBPTOut(
@@ -739,24 +750,32 @@ export class LinearPool implements PoolBase {
         try {
             // All values should use 1e18 fixed point
             // i.e. 1USDC => 1e18 not 1e6
-            const amtScaled = scale(amount, 18);
+            const amtScaled = parseFixed(amount.toString(), 18);
 
             const amt = _calcWrappedInPerBptOut(
-                amtScaled,
-                bnum(poolPairData.mainBalanceScaled.toString()),
-                bnum(poolPairData.wrappedBalanceScaled.toString()),
-                bnum(poolPairData.virtualBptSupply.toString()),
+                amtScaled.toBigInt(),
+                poolPairData.mainBalanceScaled.toBigInt(),
+                poolPairData.wrappedBalanceScaled.toBigInt(),
+                poolPairData.virtualBptSupply.toBigInt(),
                 {
-                    fee: bnum(poolPairData.swapFee.toString()),
-                    lowerTarget: bnum(poolPairData.lowerTarget.toString()),
-                    upperTarget: bnum(poolPairData.upperTarget.toString()),
+                    fee: poolPairData.swapFee.toBigInt(),
+                    lowerTarget: poolPairData.lowerTarget.toBigInt(),
+                    upperTarget: poolPairData.upperTarget.toBigInt(),
+                    rate: poolPairData.rate.toBigInt(),
                 }
             );
+
+            const amtWithRate = BigNumber.from(amt)
+                .mul(poolPairData.rate)
+                .div(ONE);
             // return human readable number
             // Using BigNumber.js decimalPlaces (dp), allows us to consider token decimal accuracy correctly,
             // i.e. when using token with 2decimals 0.002 should be returned as 0
             // Uses ROUND_DOWN mode (1)
-            return scale(amt, -18).dp(poolPairData.decimalsOut, 1);
+            return scale(bnum(amtWithRate.toString()), -18).dp(
+                poolPairData.decimalsOut,
+                1
+            );
         } catch (err) {
             return ZERO;
         }
@@ -770,24 +789,30 @@ export class LinearPool implements PoolBase {
         try {
             // All values should use 1e18 fixed point
             // i.e. 1USDC => 1e18 not 1e6
-            const amtScaled = scale(amount, 18);
+            const amtNoRate = parseFixed(amount.toString(), 18)
+                .mul(ONE)
+                .div(poolPairData.rate);
 
             const amt = _calcBptInPerWrappedOut(
-                amtScaled,
-                bnum(poolPairData.mainBalanceScaled.toString()),
-                bnum(poolPairData.wrappedBalanceScaled.toString()),
-                bnum(poolPairData.virtualBptSupply.toString()),
+                amtNoRate.toBigInt(),
+                poolPairData.mainBalanceScaled.toBigInt(),
+                poolPairData.wrappedBalanceScaled.toBigInt(),
+                poolPairData.virtualBptSupply.toBigInt(),
                 {
-                    fee: bnum(poolPairData.swapFee.toString()),
-                    lowerTarget: bnum(poolPairData.lowerTarget.toString()),
-                    upperTarget: bnum(poolPairData.upperTarget.toString()),
+                    fee: poolPairData.swapFee.toBigInt(),
+                    lowerTarget: poolPairData.lowerTarget.toBigInt(),
+                    upperTarget: poolPairData.upperTarget.toBigInt(),
+                    rate: poolPairData.rate.toBigInt(),
                 }
             );
             // return human readable number
             // Using BigNumber.js decimalPlaces (dp), allows us to consider token decimal accuracy correctly,
             // i.e. when using token with 2decimals 0.002 should be returned as 0
             // Uses ROUND_DOWN mode (1)
-            return scale(amt, -18).dp(poolPairData.decimalsOut, 1);
+            return scale(bnum(amt.toString()), -18).dp(
+                poolPairData.decimalsOut,
+                1
+            );
         } catch (err) {
             return ZERO;
         }
